controler->singleton

+++ src/main/java/Modele/Noeud.java
@@ -4,17 +4,17 @@ package Modele;
  * Created by flavi on 2017/11/18.
  */
 public class Noeud {
-    private int id;
+    private Long id;
     private int x;
     private int y;
 
-    public Noeud(int id, int x, int y) {
+    public Noeud(Long id, int x, int y) {
         this.id = id;
         this.x = x;
         this.y = y;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
